[opentitantool] No-op refactor
Break spi.rs and i2c.rs out of app/mod.rs.
