+++ sw/host/opentitanlib/BUILD
@@ -48,7 +48,9 @@ rust_library(
         "src/app/command.rs",
         "src/app/config/mod.rs",
         "src/app/config/structs.rs",
+        "src/app/i2c.rs",
         "src/app/mod.rs",
+        "src/app/spi.rs",
         "src/backend/chip_whisperer.rs",
         "src/backend/hyperdebug.rs",
         "src/backend/mod.rs",
